[Silver II] Title: 조합 0의 개수, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/2004. 조합 0의 개수/조합 0의 개수.cc
@@ -6,15 +6,15 @@ using namespace std;
 
 long long n, m;
 
-long long howmany5(long long num){
-    long long ret = 0;
+int howmany5(long long num){
+    int ret = 0;
     for(long long i=5; i<=num; i*=5)
         ret += num/i;
     return ret;
 }
 
-long long howmany2(long long num){
-    long long ret = 0;
+int howmany2(long long num){
+    int ret = 0;
     for(long long i=2; i<=num; i*=2)
         ret += num/i;
     return ret;
@@ -26,7 +26,7 @@ int main() {
 
     cin >> n >> m;
 
-    long long five, two;
+    int five, two;
     five = howmany5(n) - howmany5(m) - howmany5(n-m);
     two = howmany2(n) - howmany2(m) - howmany2(n-m);
